added data cards back and isolated Charts
